requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-docx>=1.1.0
fpdf2>=2.7.0